feat: move check to middleware
