+++ src/server.js
@@ -24,13 +24,11 @@ const logger = require('./common/utils/logger')(__filename);
 const config = require('./common/config/index');
 const availability = require('./common/config/availability');
 const router = require('./app/router');
-const noCache = require('./common/utils/no-cache');
 const autocompleteUtil = require('./common/utils/autocomplete');
 const nunjucksFilters = require('./common/utils/templateFilters.js');
 const travelPermissionCodes = require('./common/utils/travel_permission_codes.json');
 const { IS_HTTPS_SERVER, SAME_SITE_VALUE } = require('./common/config');
 const airports = require('./common/utils/airports');
-const { canSkipCsrfGeneration } = require('./common/utils/utils');
 
 // Global constants
 const PORT = process.env.PORT || 3000;
@@ -125,35 +123,6 @@ function initialiseGlobalMiddleware(app) {
       },
     })
   );
-
-  app.use((req, res, next) => {
-    res.locals.asset_path = '/public/';
-    noCache(res);
-
-    if (canSkipCsrfGeneration(req)) {
-      return next();
-    }
-
-    const token = req.csrfToken();
-    res.locals._csrf = token;
-    // This might be needed, but leaving it in for now...
-    res.cookie('XSRF-TOKEN', token, {
-      httpOnly: true,
-      secure: secureFlag,
-      sameSite: true,
-    });
-
-    // Previously, local development required the disabling of CSRF token handling
-    // The below adds the csrfToken to the res.render function which should hopefully
-    // allow for local development without this hack
-    const _render = res.render;
-    res.render = function (view, options, fn) {
-      _.extend(options, { csrfToken: token });
-      _render.call(this, view, options, fn);
-    };
-
-    return next();
-  });
 
   logger.info('Set CSRF Token');
   app.use(
